requirements.txt
gradio
transformers
torch
rdkit-pypi